Add support for firmware and tiered versions
All standard VESC BMS and VESC Express firmwares are now included by,
with an additional build option for a selection of standard bldc
firmware to be included in the VESC Tool build.

Also added packages for all different payed tiers of VESC Tool.

Updated README.md with the changes and generally rewrote large sections
of it.